--- backend/app.js
@@ -1,19 +1,0 @@
-import * as dotenv from "dotenv";
-dotenv.config();
-
-import express from "express";
-
-const app = express();
-const port = process.env.PORT || 3000;
-
-app.use(express.json());
-
-// REAL SHIT DOWN HERE
-
-app.get("/", (req, res) => {
-	res.send("Hello World!");
-});
-
-app.listen(port, () => {
-	console.log(`Example app listening on port ${port}`);
-});
